Separate receiver health borrows from synchronous world stages
